pyproject.toml
[project]
name = "garmin-turso"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "libsql-experimental>=0.0.34",
    "garminconnect>=0.2.19",
    "garth>=0.4.46",
    "schedule>=1.2.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.1",
    # Chart/Report generation
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.17.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "pandas>=2.0.0",
    "mcp>=1.14.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
